Split out classes into separate files
